reestructuré archivos y componentes, y mejoré la organización del código
