+++ package.json
@@ -2,11 +2,12 @@
   "version": "1.0.0",
   "source": "./index.html",
   "type": "module",
-  "name": "anima-project",
-  "description": "A React project automatically generated by Anima using the Shadcn UI library",
+  "name": "lincepulce",
+  "description": "Sitio Web de LincePulce",
   "scripts": {
     "dev": "vite",
-    "build": "vite build"
+    "build": "vite build",
+    "preview": "vite preview"
   },
   "dependencies": {
     "@radix-ui/react-accordion": "^1.2.1",
