+++ src/components/ui/Button.jsx
@@ -1,3 +1,4 @@
+
 const Button = ({
   variant = "primary",
   href,
@@ -7,14 +8,19 @@ const Button = ({
   children,
   ...props
 }) => {
+  // estilos base comunes para los botones
   const baseStyles = "px-5 py-2 rounded-[100px] font-medium flex items-center justify-center gap-2 transition-all duration-300 border-transparent";
+
+  // definición de variantes de estilo para el botón
   const variants = {
     primary: "bg-[#a33cfa] text-white",
     outline: "bg-[#08030d0d] text-[#08030d]",
   };
 
+  // combinación de estilos base y variantes, más clases adicionales
   const buttonClasses = `${baseStyles} ${variants[variant] || ""} ${className || ""}`;
 
+  // si se pasa un href, renderiza un enlace <a>
   if (href) {
     return (
       <a href={href} className={buttonClasses}>
@@ -25,6 +31,7 @@ const Button = ({
     );
   }
 
+  // si no hay href, renderiza un <button>
   return (
     <button className={buttonClasses} {...props}>
       {iconLeft && <span>{iconLeft}</span>}
