+++ src/components/ui/NavigationMenu.jsx
@@ -1,3 +1,5 @@
+//////////////////////////////////////////////////////////
+// Importaciones
 import * as NavigationMenuPrimitive from "@radix-ui/react-navigation-menu";
 import { cva } from "class-variance-authority";
 import { ChevronDownIcon } from "lucide-react";
@@ -5,6 +7,11 @@ import * as React from "react";
 
 import { cn } from "../../lib/utils";
 
+//////////////////////////////////////////////////////////
+
+/// forwardRef es una función que crea un componente de reenvío de referencia: https://es.reactjs.org/docs/forwarding-refs.html
+
+// Componente raíz del menú de navegación
 const NavigationMenu = React.forwardRef(({ className, children, ...props }, ref) => (
   <NavigationMenuPrimitive.Root
     ref={ref}
@@ -20,6 +27,7 @@ const NavigationMenu = React.forwardRef(({ className, children, ...props }, ref)
 ));
 NavigationMenu.displayName = NavigationMenuPrimitive.Root.displayName;
 
+// Lista de elementos en el menú de navegación
 const NavigationMenuList = React.forwardRef(({ className, ...props }, ref) => (
   <NavigationMenuPrimitive.List
     ref={ref}
@@ -32,12 +40,15 @@ const NavigationMenuList = React.forwardRef(({ className, ...props }, ref) => (
 ));
 NavigationMenuList.displayName = NavigationMenuPrimitive.List.displayName;
 
+// Item de menú, basado en el componente de Radix
 const NavigationMenuItem = NavigationMenuPrimitive.Item;
 
+// Estilo base para los triggers del menú
 const navigationMenuTriggerStyle = cva(
   "group inline-flex h-9 w-max items-center justify-center rounded-md bg-background px-4 py-2 text-sm font-medium transition-colors hover:bg-accent hover:text-accent-foreground focus:bg-accent focus:text-accent-foreground focus:outline-none disabled:pointer-events-none disabled:opacity-50 data-[active]:bg-accent/50 data-[state=open]:bg-accent/50",
 );
 
+// Botón que abre el contenido del menú
 const NavigationMenuTrigger = React.forwardRef(({ className, children, ...props }, ref) => (
   <NavigationMenuPrimitive.Trigger
     ref={ref}
@@ -53,6 +64,7 @@ const NavigationMenuTrigger = React.forwardRef(({ className, children, ...props 
 ));
 NavigationMenuTrigger.displayName = NavigationMenuPrimitive.Trigger.displayName;
 
+// Contenido del menú, aparece al hacer clic en el trigger
 const NavigationMenuContent = React.forwardRef(({ className, ...props }, ref) => (
   <NavigationMenuPrimitive.Content
     ref={ref}
@@ -65,8 +77,10 @@ const NavigationMenuContent = React.forwardRef(({ className, ...props }, ref) =>
 ));
 NavigationMenuContent.displayName = NavigationMenuPrimitive.Content.displayName;
 
+// Enlace dentro del menú de navegación
 const NavigationMenuLink = NavigationMenuPrimitive.Link;
 
+// Contenedor del menú, se asegura de que se muestre correctamente
 const NavigationMenuViewport = React.forwardRef(({ className, ...props }, ref) => (
   <div className={cn("absolute left-0 top-full flex justify-center")}>
     <NavigationMenuPrimitive.Viewport
@@ -82,6 +96,7 @@ const NavigationMenuViewport = React.forwardRef(({ className, ...props }, ref) =
 NavigationMenuViewport.displayName =
   NavigationMenuPrimitive.Viewport.displayName;
 
+// Indicador de estado del menú (por ejemplo, cuando está abierto o cerrado) 
 const NavigationMenuIndicator = React.forwardRef
   (({ className, ...props }, ref) => (
   <NavigationMenuPrimitive.Indicator
