+++ src/layout/MainLayout.jsx
@@ -1,0 +1,23 @@
+//////////////////////////////////////////////////////////
+// Importaciones
+import Header from "../components/header/Header"
+import Footer from "../components/footer/Footer"
+
+//////////////////////////////////////////////////////////
+// MainLayout es un componente que envuelve a todos los componentes de la aplicación
+/**
+ * @param {Object} props - Props del componente.
+ */
+const MainLayout = ({children}) => {
+  return (
+    <>
+      <Header />
+      <main className="w-full">
+          {children}
+      </main>
+      <Footer />
+    </>
+  )
+}
+
+export default MainLayout;
