+++ vite.config.js
@@ -11,5 +11,5 @@ export default defineConfig({
     postcss: {
       plugins: [tailwind()],
     },
-  },
-});
+  }
+})
